requirements.txt
opencv-python
deepface
dotenv
numpy
google-generativeai
python-dotenv
google-api-python-client
yt-dlp
gradio
transformers
torch
torchvision
insightface
librosa
retina-face
openai